Improve minitest integration.
First, override test method instead of hooking into the setup/teardown
hooks. This prevents missing a call/return that could occur outside of
when we start tracing. For example, when we stop tracing in
`before_teardown` we would have already seen a call event for it, but
would never see the return event.

Second, hack in something so that the typical filtered path for the
filtering recorder wouldn't reject minitest tests. I was observing
times when the return event for the test method would be coming from the
minitest gem directory instead of the test file directory.